init email_verification template: send email content in html not text
