+++ english_diary/english_diary/templates/emails/email_verification.html
@@ -1,0 +1,89 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8"><meta name="viewport" content="initial-scale=1.0"><meta name="format-detection" content="telephone=no"><title>MOSAICO Responsive Email Designer</title><style type="text/css">.socialLinks {font-size: 6px;}
+.socialLinks a {display: inline-block;}
+.socialIcon {display: inline-block;vertical-align: top;padding-bottom: 0px;border-radius: 100%;}
+table.vb-row, table.vb-content {border-collapse: separate;}
+table.vb-row {border-spacing: 9px;}
+table.vb-row.halfpad {border-spacing: 0;padding-left: 9px;padding-right: 9px;}
+table.vb-row.fullwidth {border-spacing: 0;padding: 0;}
+table.vb-container.fullwidth {padding-left: 0;padding-right: 0;}</style><style type="text/css">
+    /* yahoo, hotmail */
+    .ExternalClass, .ExternalClass p, .ExternalClass span, .ExternalClass font, .ExternalClass td, .ExternalClass div{ line-height: 100%; }
+    .yshortcuts a{ border-bottom: none !important; }
+    .vb-outer{ min-width: 0 !important; }
+    .RMsgBdy, .ExternalClass{
+      width: 100%;
+      background-color: #3f3f3f;
+      background-color: #3f3f3f}
+
+    /* outlook */
+    table{ mso-table-rspace: 0pt; mso-table-lspace: 0pt; }
+    #outlook a{ padding: 0; }
+    img{ outline: none; text-decoration: none; border: none; -ms-interpolation-mode: bicubic; }
+    a img{ border: none; }
+
+    @media screen and (max-device-width: 600px), screen and (max-width: 600px) {
+      table.vb-container, table.vb-row{
+        width: 95% !important;
+      }
+
+      .mobile-hide{ display: none !important; }
+      .mobile-textcenter{ text-align: center !important; }
+
+      .mobile-full{
+        float: none !important;
+        width: 100% !important;
+        max-width: none !important;
+        padding-right: 0 !important;
+        padding-left: 0 !important;
+      }
+      img.mobile-full{
+        width: 100% !important;
+        max-width: none !important;
+        height: auto !important;
+      }   
+    }
+  </style><style type="text/css">#ko_textBlock_3 .links-color a, #ko_textBlock_3 .links-color a:link, #ko_textBlock_3 .links-color a:visited, #ko_textBlock_3 .links-color a:hover {color: #3f3f3f;color: #3f3f3f;text-decoration: underline;}
+#ko_textBlock_3 .long-text p:first-child {margin-top: 0px;}
+#ko_footerBlock_2 .links-color a:visited, #ko_footerBlock_2 .links-color a:hover {color: #ccc;color: #ccc;text-decoration: underline;}</style></head><body bgcolor="#3f3f3f" text="#919191" alink="#cccccc" vlink="#cccccc" style="margin: 0;padding: 0;background-color: #3f3f3f;color: #919191;">
+
+  <center>
+
+  <!-- preheaderBlock -->
+  
+  <!-- /preheaderBlock -->
+
+  <table class="vb-outer" width="100%" cellpadding="0" border="0" cellspacing="0" bgcolor="#bfbfbf" style="background-color: #bfbfbf;" id="ko_textBlock_3"><tbody><tr><td class="vb-outer" align="center" valign="top" bgcolor="#bfbfbf" style="padding-left: 9px;padding-right: 9px;background-color: #bfbfbf;">
+
+<!--[if (gte mso 9)|(lte ie 8)]><table align="center" border="0" cellspacing="0" cellpadding="0" width="570"><tr><td align="center" valign="top"><![endif]-->
+        <div class="oldwebkit" style="max-width: 570px;">
+        <table width="570" border="0" cellpadding="0" cellspacing="18" class="vb-container fullpad" bgcolor="#ffffff" style="border-collapse: separate;border-spacing: 18px;padding-left: 0;padding-right: 0;width: 100%;max-width: 570px;background-color: #fff;"><tbody><tr><td align="left" class="long-text links-color" style="text-align: left; font-size: 13px; font-family: Arial, Helvetica, sans-serif; color: #3f3f3f;"><h2><strong>깐깐하게 쓰는 영어일기</strong></h2><p style="margin: 1em 0px;"><br data-mce-bogus="1"></p><p style="margin: 1em 0px;">안녕하세요, {{ username }}님.</p><p style="margin: 1em 0px;">가입을 완료하려면 먼저 이메일 주소를 인증해야 합니다.</p><p style="margin: 1em 0px;margin-bottom: 0px;"><br data-mce-bogus="1"></p></td>
+          </tr></tbody></table></div>
+<!--[if (gte mso 9)|(lte ie 8)]></td></tr></table><![endif]-->
+      </td>
+    </tr></tbody></table><table class="vb-outer" width="100%" cellpadding="0" border="0" cellspacing="0" bgcolor="#bfbfbf" style="background-color: #bfbfbf;" id="ko_buttonBlock_4"><tbody><tr><td class="vb-outer" align="center" valign="top" bgcolor="#bfbfbf" style="padding-left: 9px;padding-right: 9px;background-color: #bfbfbf;">
+
+<!--[if (gte mso 9)|(lte ie 8)]><table align="center" border="0" cellspacing="0" cellpadding="0" width="570"><tr><td align="center" valign="top"><![endif]-->
+        <div class="oldwebkit" style="max-width: 570px;">
+        <table width="570" border="0" cellpadding="0" cellspacing="18" class="vb-container fullpad" bgcolor="#ffffff" style="border-collapse: separate;border-spacing: 18px;padding-left: 0;padding-right: 0;width: 100%;max-width: 570px;background-color: #fff;"><tbody><tr><td valign="top" bgcolor="#ffffff" align="center" style="background-color: #ffffff;">
+
+              <table cellpadding="0" border="0" align="center" cellspacing="0" class="mobile-full"><tbody><tr><td width="auto" valign="middle" bgcolor="#e53935" align="center" height="50" style="font-size: 15px; font-family: Arial, Helvetica, sans-serif; color: #fff; font-weight: normal; padding-left: 14px; padding-right: 14px; background-color: #e53935; border-radius: 4px;">
+                    <a target="_new" href="https://engdiary.dsds101.info/" style="text-decoration: none; color: #fff; font-weight: normal;">이메일 인증하기</a>
+                  </td>
+                </tr></tbody></table></td>
+          </tr></tbody></table></div>
+<!--[if (gte mso 9)|(lte ie 8)]></td></tr></table><![endif]-->
+      </td>
+    </tr></tbody></table><table class="vb-outer" width="100%" cellpadding="0" border="0" cellspacing="0" bgcolor="#bfbfbf" style="background-color: #bfbfbf;" id="ko_titleBlock_7"><tbody><tr><td class="vb-outer" align="center" valign="top" bgcolor="#bfbfbf" style="padding-left: 9px;padding-right: 9px;background-color: #bfbfbf;">
+
+<!--[if (gte mso 9)|(lte ie 8)]><table align="center" border="0" cellspacing="0" cellpadding="0" width="570"><tr><td align="center" valign="top"><![endif]-->
+        <div class="oldwebkit" style="max-width: 570px;">
+        <table width="570" border="0" cellpadding="0" cellspacing="9" class="vb-container halfpad" bgcolor="#ffffff" style="border-collapse: separate;border-spacing: 9px;padding-left: 9px;padding-right: 9px;width: 100%;max-width: 570px;background-color: #fff;"><tbody><tr><td bgcolor="#ffffff" align="left" style="background-color: #ffffff; font-size: 20px; font-family: Arial, Helvetica, sans-serif; color: #3f3f3f; text-align: left;">
+              <span></span>
+            </td>
+          </tr></tbody></table></div>
+<!--[if (gte mso 9)|(lte ie 8)]></td></tr></table><![endif]-->
+      </td>
+    </tr></tbody></table><!-- footerBlock --><!-- /footerBlock --></center>
+
+</body></html>
